Adding an option to bypass pools entirely
In some cases, a few high priority requests happening with
a low frequency may be seen as admissible -- for example,
important error messages, diagnostic polls, and so on.

This change adds the possibility set the option 'max_connections'
to {max_connections, bypass} to have a socket opened that does
not respect a given's Scheme+Host+Port's pool limits and opens
a socket outside of it.

This socket will be closed as soon as the response is obtained.